add as methods on eventsClient
